Implement and use mock of Chat Channels
